cleanup crap
