--- controller/test/e2e/manifests/controlplane-plaintext-helm.yaml
@@ -1,3 +1,0 @@
-controller:
-  xds:
-    mode: plaintext
