--- controller/test/e2e/manifests/controlplane-tls-helm.yaml
@@ -1,3 +1,0 @@
-controller:
-  xds:
-    mode: tls
