--- controller/test/e2e/manifests/custom-gwp-2.yaml
@@ -1,7 +1,0 @@
-gatewayClassParametersRefs:
-  kgateway:
-    name: custom-gwp-2
-    namespace: kgateway-test
-  agentgateway:
-    name: custom-agwp-2
-    namespace: kgateway-test
