--- controller/test/e2e/manifests/inference-extension-helm.yaml
@@ -1,10 +1,0 @@
-# Remove image registry directive when https://github.com/kgateway-dev/kgateway/issues/11359 is fixed
-image:
-  registry: ghcr.io/agentgateway
-
-# Configuration for the agentgateway controller pod
-controller:
-  logLevel: debug
-
-inferenceExtension:
-  enabled: true
